+++ src/components/content/exercise.tsx
@@ -116,6 +116,8 @@ export function Exercise({ node }: ExerciseProps) {
 
     const state = node.taskEdtrState
 
+    console.log('render interactive', node)
+
     if (state.interactive) {
       if (state.interactive.plugin === 'scMcExercise') {
         return (
