+++ src/components/content/sc-mc-exercise.tsx
@@ -16,6 +16,7 @@ export interface ScMcExerciseProps {
 }
 
 export function ScMcExercise({ state, idBase }: ScMcExerciseProps) {
+  console.log(idBase)
   if (state.isSingleChoice)
     return <SingleChoice state={state} idBase={idBase} />
